requirements.txt
asgiref
attrs
dj-database-url
Django
django-cors-headers
djangorestframework
drf-spectacular
gunicorn
inflection
jsonschema
jsonschema-specifications
packaging
pillow
psycopg2-binary
PyYAML
referencing
rpds-py
sqlparse
typing_extensions
tzdata
uritemplate
whitenoise
